requirements.txt
langchain==0.0.257
openai~=0.28.0


requests~=2.31.0
websocket-client==1.6.3